[#10785] work around matplotlib warning caused by PyInstaller
The warning is caused by PyInstaller's runtime hook for matplotlib, and
prevents the Astropy tests from running without error when testing the
PyInstaller build in our CI.

This workaround does not fix the issue in general: e.g. users using PyInstaller
to bundle an application that users Astropy will have to include a similar fix
in their own code (as the issue is not unique to Astropy).