Ensure Field can access VOTableFile version flags
Because the unit format of a VOTable depends on its version the Field
class must have access to this information, including in the test
functions. This commit adds a function to VOTableFile that creates the
appropriate Boolean flags. The same function can also be used in the
VOTableFile parse() and to_xml() methods.