[#11094] Remove use of OrderedDescriptor for astropy.modeling
Parameter.__set_name__ is used to set the names of Parameters upon class
initialization.

As before, the Model._parameters_ dict is intialized in
_ModelMeta.__new__ (which can't easily be replaced with
__init_subclass__ as it has other functionalities).

Although the removed function astropy.utils.misc.get_parameters was
technically public, it was arguably an implementation detail of
astropy.modeling.core and never should have been there; it is easily
replaced now with a one-liner.