Remove lock from lazyproperty setter+deleter
If the user is calling these concurrently from multiple threads, they
already have a race condition in their code (as opposed to using the
getter from multiple threads, which they can have a reasonable
expectation of working).

I did change the deleter to do the deletion atomically, so that as long
as no fdel is provided it'll be safe to do concurrent deletion. This
should restore performance for classes that use a lazyproperty as a
cache and need to invalidate the cache when attributes change (e.g.
Time).