Refactor physical types
Let def_physical_type expand definitions of physical types

This change keeps backwards compatibility in mind.  Suppose someone
has a package in which they define a physical type for a unit that
hasn't been defined already in Astropy.  Then suppose a future version
of Astropy defines that unit.  Before now, this would cause an
exception saying that the physical type has already been defined.
After now, def_physical_type will not fail, but rather add a new name
to the unit (so long as that physical type name is not already in
use).