Refactor physical types
Update docstrings for physical types.

I clarified that the physical types are defined as being dimensionally
compatible with each other, given the example that pascals are
intended for pressure, but are dimensionally equivalent to energy
density.  I also edited the physical type names for ohms.

Allow physical types to be multiplied/divided by numbers

For example, multiplying a physical type by a real number will return the
original physical type.  The problem I had been running into was that if I
multiplied a Unit (on the left) with a PhysicalType (on the right), then
the operation would return a Quantity.  Having __array_ufunc__ raise a
TypeError prevents this from happening...somehow.

Refactor tests for physical types (including operations)

I previously had a bunch of individual tests that separately tested division,
multiplication, exponentiation, equality, etc.  This did not end up scaling
well when I wanted to add new tests.  I parametrized the tests of the dunder
methods for PhysicalType objects to make it much easier to add new test cases.
I was able to consolidate a whole bunch of test cases that way.

I also added a test to make sure that PhysicalType objects cannot case into
Quantity objects, since that messed up several of the dunder operations when
an object besides a PhysicalType was on the left.

Remove PhysicalType.__array_ufunc__ TypeError

Apparently having this method is enough to prevent PhysicalType objects from
being cast as Quantity objects.  It's not necessary to raise a TypeError.
I do not have a good understanding why.